requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-filters==0.2.1
django-oauth-toolkit==1.0.0
djangorestframework==3.7.3
idna==2.6
mysqlclient==1.3.12
oauthlib==2.0.6
pytz==2017.3
requests==2.18.4
rest-condition==1.0.3
urllib3==1.22
